impl step_03

+++ src/asm/mod.rs
@@ -108,7 +108,7 @@ impl fmt::Display for Ins {
 #[allow(dead_code)]
 pub enum Opr {
     Direct(Reg),
-    Literal(u32),
+    Literal(u64),
 }
 
 impl fmt::Display for Opr {
